feat(configuration): merging joints configuration panels
